+++ project.go
@@ -46,10 +46,10 @@ type ProjectList struct {
 
 // Project represents a Terraform Enterprise project
 type Project struct {
-	ID   string `jsonapi:"primary,projects"`
-	Name string `jsonapi:"attr,name"`
+	ID        string `jsonapi:"primary,projects"`
+	IsUnified bool   `jsonapi:"attr,is-unified"`
+	Name      string `jsonapi:"attr,name"`
 
-	// **Note: This field is still in BETA and subject to change.**
 	Description string `jsonapi:"attr,description"`
 
 	// Relations
@@ -81,7 +81,6 @@ type ProjectCreateOptions struct {
 	Name string `jsonapi:"attr,name"`
 
 	// Optional: A description for the project.
-	// **Note: This field is still in BETA and subject to change.**
 	Description *string `jsonapi:"attr,description,omitempty"`
 }
 
@@ -97,7 +96,6 @@ type ProjectUpdateOptions struct {
 	Name *string `jsonapi:"attr,name,omitempty"`
 
 	// Optional: A description for the project.
-	// **Note: This field is still in BETA and subject to change.**
 	Description *string `jsonapi:"attr,description,omitempty"`
 }